♻️ Refactor : 자료받아오는 방식 수정 검색컴포넌트  코드 변수명 수정
이해하기 쉬운 변수명으로 수정
